packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.11.0",
    "diffusers>=0.35.2",
    "edge-tts>=7.2.3",
    "fastapi>=0.121.2",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
]


[tool.uv]
package = true

[project.scripts]
test = "api.main:run_tests"
serve = "api.main:run_server"
dev = "api.main:run_dev"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
